+++ datadog-trace-protobuf/src/remoteconfig.rs
@@ -3,8 +3,7 @@
 
 use serde::{Deserialize, Serialize};
 // This file is @generated by prost-build.
-#[derive(Deserialize, Serialize)]
-#[derive(Clone, PartialEq, ::prost::Message)]
+#[derive(Deserialize, Serialize, Clone, PartialEq, ::prost::Message)]
 pub struct File {
     #[prost(string, tag = "1")]
     pub path: ::prost::alloc::string::String,
@@ -12,8 +11,7 @@ pub struct File {
     #[serde(with = "serde_bytes")]
     pub raw: ::prost::alloc::vec::Vec<u8>,
 }
-#[derive(Deserialize, Serialize)]
-#[derive(Clone, PartialEq, ::prost::Message)]
+#[derive(Deserialize, Serialize, Clone, PartialEq, ::prost::Message)]
 pub struct Client {
     #[prost(message, optional, tag = "1")]
     pub state: ::core::option::Option<ClientState>,
@@ -35,8 +33,7 @@ pub struct Client {
     #[prost(bytes = "vec", tag = "11")]
     pub capabilities: ::prost::alloc::vec::Vec<u8>,
 }
-#[derive(Deserialize, Serialize)]
-#[derive(Clone, PartialEq, ::prost::Message)]
+#[derive(Deserialize, Serialize, Clone, PartialEq, ::prost::Message)]
 pub struct ClientTracer {
     #[prost(string, tag = "1")]
     pub runtime_id: ::prost::alloc::string::String,
@@ -55,8 +52,7 @@ pub struct ClientTracer {
     #[prost(string, repeated, tag = "7")]
     pub tags: ::prost::alloc::vec::Vec<::prost::alloc::string::String>,
 }
-#[derive(Deserialize, Serialize)]
-#[derive(Clone, PartialEq, ::prost::Message)]
+#[derive(Deserialize, Serialize, Clone, PartialEq, ::prost::Message)]
 pub struct ClientAgent {
     #[prost(string, tag = "1")]
     pub name: ::prost::alloc::string::String,
@@ -69,8 +65,7 @@ pub struct ClientAgent {
     #[prost(string, repeated, tag = "5")]
     pub cws_workloads: ::prost::alloc::vec::Vec<::prost::alloc::string::String>,
 }
-#[derive(Deserialize, Serialize)]
-#[derive(Clone, PartialEq, ::prost::Message)]
+#[derive(Deserialize, Serialize, Clone, PartialEq, ::prost::Message)]
 pub struct ConfigState {
     #[prost(string, tag = "1")]
     pub id: ::prost::alloc::string::String,
@@ -83,8 +78,7 @@ pub struct ConfigState {
     #[prost(string, tag = "5")]
     pub apply_error: ::prost::alloc::string::String,
 }
-#[derive(Deserialize, Serialize)]
-#[derive(Clone, PartialEq, ::prost::Message)]
+#[derive(Deserialize, Serialize, Clone, PartialEq, ::prost::Message)]
 pub struct ClientState {
     #[prost(uint64, tag = "1")]
     pub root_version: u64,
@@ -99,16 +93,14 @@ pub struct ClientState {
     #[prost(bytes = "vec", tag = "6")]
     pub backend_client_state: ::prost::alloc::vec::Vec<u8>,
 }
-#[derive(Deserialize, Serialize)]
-#[derive(Clone, PartialEq, ::prost::Message)]
+#[derive(Deserialize, Serialize, Clone, PartialEq, ::prost::Message)]
 pub struct TargetFileHash {
     #[prost(string, tag = "1")]
     pub algorithm: ::prost::alloc::string::String,
     #[prost(string, tag = "3")]
     pub hash: ::prost::alloc::string::String,
 }
-#[derive(Deserialize, Serialize)]
-#[derive(Clone, PartialEq, ::prost::Message)]
+#[derive(Deserialize, Serialize, Clone, PartialEq, ::prost::Message)]
 pub struct TargetFileMeta {
     #[prost(string, tag = "1")]
     pub path: ::prost::alloc::string::String,
@@ -117,16 +109,14 @@ pub struct TargetFileMeta {
     #[prost(message, repeated, tag = "3")]
     pub hashes: ::prost::alloc::vec::Vec<TargetFileHash>,
 }
-#[derive(Deserialize, Serialize)]
-#[derive(Clone, PartialEq, ::prost::Message)]
+#[derive(Deserialize, Serialize, Clone, PartialEq, ::prost::Message)]
 pub struct ClientGetConfigsRequest {
     #[prost(message, optional, tag = "1")]
     pub client: ::core::option::Option<Client>,
     #[prost(message, repeated, tag = "2")]
     pub cached_target_files: ::prost::alloc::vec::Vec<TargetFileMeta>,
 }
-#[derive(Deserialize, Serialize)]
-#[derive(Clone, PartialEq, ::prost::Message)]
+#[derive(Deserialize, Serialize, Clone, PartialEq, ::prost::Message)]
 pub struct ClientGetConfigsResponse {
     #[prost(bytes = "vec", repeated, tag = "1")]
     #[serde(with = "crate::serde")]
